+++ pos_product_available/static/src/js/pos.js
@@ -12,9 +12,9 @@ odoo.define('pos_product_available.PosModel', function(require){
 
             var loaded = PosModelSuper.load_server_data.call(this);
 
-            var set_prod_qtys = function(products) {
-                _.each(products, function(pr){
-                    self.db.get_product_by_id(pr.id).qty_available = pr.qty_available;
+            var set_prod_vals = function(vals) {
+                _.each(vals, function(v){
+                    _.extend(self.db.get_product_by_id(v.id), v);
                 });
             };
 
@@ -22,7 +22,7 @@ odoo.define('pos_product_available.PosModel', function(require){
                 return model.model === 'product.product';
             });
             if (prod_model) {
-                prod_model.fields.push('qty_available');
+                prod_model.fields.push('qty_available', 'type');
                 var context_super = prod_model.context;
                 prod_model.context = function(that){
                     var ret = context_super(that);
@@ -32,16 +32,16 @@ odoo.define('pos_product_available.PosModel', function(require){
                 var loaded_super = prod_model.loaded;
                 prod_model.loaded = function(that, products){
                     loaded_super(that, products);
-                    set_prod_qtys(products);
+                    set_prod_vals(products);
                 };
                 return loaded;
             }
 
             return loaded.then(function(){
-                return new Model('product.product').query(['qty_available']).
+                return new Model('product.product').query(['qty_available', 'type']).
                     filter([['sale_ok','=',true],['available_in_pos','=',true]]).
                     context({'location': self.config.stock_location_id[0]}).all().then(function(products){
-                        set_prod_qtys(products);
+                        set_prod_vals(products);
                     });
             });
         },